refactor(encoding): extract price-set SIMD kernels to hc/priceset.rs
Move the optimal parser's per-CPU-tier price-set kernels
(priceset_range_nonabort_{scalar,avx2,neon,sse41,simd128} plus their
cached-price / improved-mask helpers) out of match_generator.rs into
hc/priceset.rs. Verbatim move (kernels unchanged; only the opt-parser type
imports pulled in and the nonabort entry points widened to pub(crate)), so the
output is byte-identical. match_generator.rs: 12006 -> 11306.

Part of #441.
